- Turned double volume variables into float - Macro names turned into upper-case

+++ trunk/src/vba/apu/Gb_Apu.h
@@ -86,7 +86,7 @@ class Gb_Apu
 	// Sound adjustments
 
 	// Sets overall volume, where 1.0 is normal.
-	void volume( double );
+	void volume( float );
 
 	// If true, reduces clicking by disabling DAC biasing. Note that this reduces
 	// emulation accuracy, since the clicks are authentic.
@@ -109,7 +109,7 @@ class Gb_Apu
 	Gb_Osc*		oscs [OSC_COUNT];
 	int32_t		last_time;          // time sound emulator has been run to
 	int32_t		frame_period;       // clocks between each frame sequencer step
-	double		volume_;
+	float		volume_;
 	bool		reduce_clicks_;
 
 	Gb_Sweep_Square square1;
